refactor(input-icon-toggle): make experimental input-icon-toggle component internal
BREAKING CHANGE: New component path src/__internal__/input-icon-toggle
