+++ src/__experimental__/components/textarea/textarea.component.js
@@ -8,7 +8,7 @@ import CharacterCount from "./character-count";
 import Input from "../input/input.component";
 import { InputBehaviour } from "../../../__internal__/input-behaviour";
 
-import InputIconToggle from "../input-icon-toggle";
+import InputIconToggle from "../../../__internal__/input-icon-toggle";
 
 import guid from "../../../utils/helpers/guid/guid";
 import StyledTextarea from "./textarea.style";
