+++ src/__internal__/input-icon-toggle/index.js
