+++ src/__internal__/input-icon-toggle/input-icon-toggle.component.js
@@ -1,9 +1,9 @@
 import React from "react";
 import PropTypes from "prop-types";
-import Icon from "../../../components/icon";
+import Icon from "../../components/icon";
 import InputIconToggleStyle from "./input-icon-toggle.style";
-import OptionsHelper from "../../../utils/helpers/options-helper";
-import ValidationIcon from "../../../components/validations/validation-icon.component";
+import OptionsHelper from "../../utils/helpers/options-helper/options-helper";
+import ValidationIcon from "../../components/validations/validation-icon.component";
 
 const shouldDisplayValidationIcon = ({ error, warning, info }) => {
   const validation = error || warning || info || null;
