+++ src/components/select/select.style.js
@@ -3,7 +3,7 @@ import { margin } from "styled-system";
 
 import InputPresentationStyle from "../../__experimental__/components/input/input-presentation.style";
 import StyledInput from "../../__experimental__/components/input/input.style";
-import InputIconToggleStyle from "../../__experimental__/components/input-icon-toggle/input-icon-toggle.style";
+import InputIconToggleStyle from "../../__internal__/input-icon-toggle/input-icon-toggle.style";
 import { baseTheme } from "../../style/themes";
 
 const StyledSelect = styled.div`
