+++ src/components/select/simple-select/simple-select.spec.js
@@ -11,7 +11,7 @@ import Textbox from "../../../__experimental__/components/textbox";
 import Option from "../option/option.component";
 import SelectList from "../select-list/select-list.component";
 import SelectTextbox from "../select-textbox/select-textbox.component";
-import InputIconToggleStyle from "../../../__experimental__/components/input-icon-toggle/input-icon-toggle.style";
+import InputIconToggleStyle from "../../../__internal__/input-icon-toggle/input-icon-toggle.style";
 import StyledInput from "../../../__experimental__/components/input/input.style";
 import InputPresentationStyle from "../../../__experimental__/components/input/input-presentation.style";
 import { baseTheme } from "../../../style/themes";
